Create the .bash_untracked file if it doesn't exist

+++ install.sh
@@ -1,6 +1,6 @@
 DIR=~/dotfiles # dotfiles directory
 OLDDIR=~/dotfiles_old # old dotfiles backup directory
-# list of files/folders to symlink in homedir
+# list of files to symlink in homedir
 FILES="ackrc bashrc vimrc bash_aliases bash_functions tmux.conf pylintrc"
 
 mkdir -p $OLDDIR
@@ -8,20 +8,29 @@ cd $DIR
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create
 # symlinks from the homedir to any files in the ~/dotfiles directory specified
-# in $files
+# in $FILES
 for file in $FILES; do
-    mv ~/.$file $OLDDIR/
+    if [ -f ~/.$file ] ; then
+        mv ~/.$file $OLDDIR/
+    fi
     echo "Creating symlink to $file in home directory."
     ln -s $DIR/$file ~/.$file
 done
+
+if [ ! -f ~/.bash_untracked ] ; then
+    cat << EOF > ~/.bash_untracked
+# Use this file to store system specific aliases and functions
+# It will not be versioned, and anything declared in here will
+# override the other values
+EOF
+fi
 
 echo "Installing missing any system dependencies - ^C to cancel"
 if hash yum 2>/dev/null; then
     sudo yum install ctags python-virtualenvwrapper tmux git ack pylint nodejs
 else
-    sudo apt-get install exuberant-ctags tmux git pyflakes
+    sudo apt-get install exuberant-ctags virtualenvwrapper tmux git pyflakes nodejs
 fi
-
 sudo npm install -g jshint
 
 echo "Installing VIM bundles"
